// test/src/connectors/kujira/kujira.types.ts
export type IMap<K, V> = ImmutableMap<K, V>;
export const IMap = ImmutableMap;

export type BasicKujiraToken = Denom;
export type BasicKujiraMarket = fin.Pair;

export type KujiraWithdraw = ExecuteResult;

export type Address = string;
export type OwnerAddress = Address;
export type PayerAddress = Address;
export type Price = BigNumber;
export type Amount = BigNumber;
export type Fee = BigNumber;
export type Timestamp = number;
export type Block = number;
export type EncryptedWallet = string;

export type ConnectorMarket = any;
export type ConnectorTicker = any;
export type ConnectorOrderBook = any;
export type ConnectorOrder = any;

export type TokenId = Address;
export type TokenName = string;
export type TokenSymbol = string;
export type TokenDecimals = number;

export type MarketName = string;
export type MarketId = Address;
export type MarketPrecision = number;
export type MarketProgramId = Address;
export type MarketDeprecation = boolean;
export type MarketMinimumOrderSize = BigNumber;
export type MarketMinimumPriceIncrement = BigNumber;
export type MarketMinimumBaseIncrement = BigNumber;
export type MarketMinimumQuoteIncrement = BigNumber;

export type TickerPrice = Price;
export type TickerTimestamp = Timestamp;

export type TransactionHash = string;

export type OrderId = string;
export type OrderClientId = string;
export type OrderMarketName = MarketName;
export type OrderMarketId = MarketId;
export type OrderMarket = Market;
export type OrderOwnerAddress = OwnerAddress;
export type OrderPayerAddress = PayerAddress;
export type OrderPrice = Price;
export type OrderAmount = Amount;
export type OrderFee = Fee;
export type OrderCreationTimestamp = Timestamp;
export type OrderFillingTimestamp = Timestamp;
export type OrderTransactionHashes = TransactionHashes;
// ...
export enum OrderSide {
  BUY = 'BUY',
  SELL = 'SELL',
}

export enum OrderStatus {
  OPEN = 'OPEN',
  CANCELLED = 'CANCELLED',
  PARTIALLY_FILLED = 'PARTIALLY_FILLED',
  FILLED = 'FILLED',
  CREATION_PENDING = 'CREATION_PENDING',
  CANCELLATION_PENDING = 'CANCELLATION_PENDING',
  UNKNOWN = 'UNKNOWN',
}

export enum OrderType {
  MARKET = 'MARKET',
  LIMIT = 'LIMIT',
  IOC = 'IOC', // Immediate or Cancel
}
// ...
export interface Token {
  id: TokenId;
  name: TokenName;
  symbol: TokenSymbol;

  decimals: TokenDecimals;
}

export interface Market {
  id: MarketId;
  name: MarketName;
  baseToken: Token;
  quoteToken: Token;
  precision: MarketPrecision;
  minimumOrderSize: MarketMinimumOrderSize;
  minimumPriceIncrement: MarketMinimumPriceIncrement; // Also called tickSize
  minimumBaseAmountIncrement: MarketMinimumBaseIncrement;
  minimumQuoteAmountIncrement: MarketMinimumQuoteIncrement;
  fees: MarketFee;
  programId?: MarketProgramId;
  deprecated?: MarketDeprecation;
  connectorMarket: ConnectorMarket;
}
// ...
export interface Order {
  id?: OrderId;
  clientId?: OrderClientId; // Client custom id
  marketName: OrderMarketName;
  marketId: OrderMarketId;
  market: OrderMarket;
  ownerAddress?: OrderOwnerAddress;
  payerAddress?: OrderPayerAddress;
  price?: OrderPrice;
  amount: OrderAmount;
  side: OrderSide;
  status?: OrderStatus;
  type?: OrderType;
  fee?: OrderFee;
  creationTimestamp?: OrderCreationTimestamp;
  fillingTimestamp?: OrderFillingTimestamp;
  hashes?: OrderTransactionHashes;
  connectorOrder?: ConnectorOrder;
}

export interface TransactionHashes {
  creation?: TransactionHash;
  cancellation?: TransactionHash;
  withdraw?: TransactionHash;
  creations?: TransactionHash[];
  cancellations?: TransactionHash[];
  withdraws?: TransactionHash[];
}

export interface MarketFee {
  maker: FeeMaker;
  taker: FeeTaker;
  serviceProvider: FeeServiceProvider;
}
// ...
export interface GetAllMarketsResponse extends IMap<MarketId, Market> {}
// ...
export type CancelOrdersResponse =
  | IMap<OrderId, Order>
  | IMap<OwnerAddress, IMap<OrderId, Order>>;